[FEDEXSHIP-31] moved advanced configuration to new file
